requirements.txt
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
python-dateutil>=2.8.2